# revenue-share-platform/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Reven — Partner Revenue OS API
  version: 1.0.0
  description: >
    Phase 1 "Capture" API for partner revenue and revenue-sharing governance:
    partners, contracts, governed amendment journeys, revenue-share records,
    KPIs, legal-document metadata, and dashboard analytics.


    Phase boundary: this API does NOT move money. It records and previews
    revenue-share data and payout-readiness status only. There is no
    payment-execution, settlement, tax-clearance, or ERP-write endpoint.
servers:
  - url: /api
    description: Same-origin API base path
...
security:
  - bearerAuth: []
paths:
...
  /partners:
    get:
      tags: [Partners]
      summary: List partners
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: type, schema: { type: string } }
      responses:
        '200':
          description: Partners
          content:
            application/json:
              schema:
                type: object
                properties:
                  partners:
                    type: array
                    items: { $ref: '#/components/schemas/Partner' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Partners]
      summary: Create a partner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  partner: { $ref: '#/components/schemas/Partner' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /contracts:
    get:
      tags: [Contracts]
      summary: List contracts
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: partnerId, schema: { type: integer } }
      responses:
        '200':
          description: Contracts
          content:
            application/json:
              schema:
                type: object
                properties:
                  contracts:
                    type: array
                    items: { $ref: '#/components/schemas/Contract' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Contracts]
      summary: Create a contract
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContractInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  contract: { $ref: '#/components/schemas/Contract' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /amendments:
    get:
      tags: [Amendments]
      summary: List amendment journeys
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: contractId, schema: { type: integer } }
      responses:
        '200':
          description: Amendment journeys (each includes a readiness object)
          content:
            application/json:
              schema:
                type: object
                properties:
                  amendments:
                    type: array
                    items: { $ref: '#/components/schemas/Amendment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Amendments]
      summary: Create an amendment journey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AmendmentInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  amendment: { $ref: '#/components/schemas/Amendment' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /revenue:
    get:
      tags: [Revenue]
      summary: List revenue-share records
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: contractId, schema: { type: integer } }
      responses:
        '200':
          description: Revenue-share records
          content:
            application/json:
              schema:
                type: object
                properties:
                  revenueShares:
                    type: array
                    items: { $ref: '#/components/schemas/RevenueShare' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Revenue]
      summary: Create a revenue-share record (records/previews only — no payment)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RevenueShareInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  revenueShare: { $ref: '#/components/schemas/RevenueShare' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /kpis:
    get:
      tags: [KPIs]
      summary: List KPIs
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: contractId, schema: { type: integer } }
      responses:
        '200':
          description: KPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  kpis:
                    type: array
                    items: { $ref: '#/components/schemas/KPI' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [KPIs]
      summary: Create a KPI
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KPIInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  kpi: { $ref: '#/components/schemas/KPI' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /documents:
    get:
      tags: [Documents]
      summary: List legal-document metadata
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: contractId, schema: { type: integer } }
        - { in: query, name: documentType, schema: { type: string } }
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items: { $ref: '#/components/schemas/LegalDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Documents]
      summary: Create a legal-document metadata record
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LegalDocumentInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  document: { $ref: '#/components/schemas/LegalDocument' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /claims:
    get:
      tags: [Claims]
      summary: List partner revenue claims
      parameters:
        - { in: query, name: status, schema: { type: string } }
        - { in: query, name: partnerId, schema: { type: integer } }
        - { in: query, name: contractId, schema: { type: integer } }
      responses:
        '200':
          description: Claims
          content:
            application/json:
              schema:
                type: object
                properties:
                  claims:
                    type: array
                    items: { $ref: '#/components/schemas/Claim' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Claims]
      summary: Submit a partner revenue claim
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimInput' }
      responses:
        '201':
          description: Submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  claim: { $ref: '#/components/schemas/Claim' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /audit:
    get:
      tags: [Audit]
      summary: List audit events (append-only history)
      parameters:
        - { in: query, name: entityType, schema: { type: string } }
        - { in: query, name: entityId, schema: { type: integer } }
        - { in: query, name: actorUserId, schema: { type: integer } }
      responses:
        '200':
          description: Audit events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Missing or invalid fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Partner:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        company: { type: string }
        type: { type: string }
        status: { type: string }
        contact_person: { type: string }
        phone: { type: string }
        address: { type: string }
        notes: { type: string }
    PartnerInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        company: { type: string }
        type: { type: string }
        status: { type: string }
        contactPerson: { type: string }
        phone: { type: string }
        address: { type: string }
        notes: { type: string }
    Contract:
      type: object
      properties:
        id: { type: integer }
        partner_id: { type: integer }
        partner_name: { type: string }
        title: { type: string }
        description: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        revenue_share_percentage: { type: number }
        minimum_payout: { type: number }
        payment_terms: { type: string }
        status: { type: string }
    ContractInput:
      type: object
      required: [partnerId, title, startDate, revenueSharePercentage]
      properties:
        partnerId: { type: integer }
        title: { type: string }
        description: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        revenueSharePercentage: { type: number }
        minimumPayout: { type: number }
        paymentTerms: { type: string }
        status: { type: string }
    Readiness:
      type: object
      properties:
        ready: { type: boolean }
        missing: { type: array, items: { type: string } }
        total: { type: integer }
        satisfied: { type: integer }
    Amendment:
      type: object
      properties:
        id: { type: integer }
        contract_id: { type: integer }
        contract_title: { type: string }
        article_reference: { type: string }
        amendment_type: { type: string }
        amendment_mechanism: { type: string }
        reason: { type: string }
        public_interest_basis: { type: string }
        necessity_confirmed: { type: boolean }
        no_new_contract_confirmed: { type: boolean }
        no_nature_change_confirmed: { type: boolean }
        notice_period_days: { type: integer }
        authority_source: { type: string }
        decision_date: { type: string, format: date }
        effective_date: { type: string, format: date }
        partner_impact: { type: string }
        calculation_method: { type: string }
        amendment_letter_reference: { type: string }
        notice_channels: { type: array, items: { type: string } }
        notice_message: { type: string }
        notice_message_language: { type: string }
        status: { type: string, enum: [draft, ready, notified, acknowledged] }
        readiness: { $ref: '#/components/schemas/Readiness' }
    AmendmentInput:
      type: object
      required: [contractId]
      properties:
        contractId: { type: integer }
        articleReference: { type: string }
        amendmentType: { type: string }
        amendmentMechanism: { type: string }
        reason: { type: string }
        publicInterestBasis: { type: string }
        necessityConfirmed: { type: boolean }
        noNewContractConfirmed: { type: boolean }
        noNatureChangeConfirmed: { type: boolean }
        noticePeriodDays: { type: integer }
        authoritySource: { type: string }
        decisionDate: { type: string, format: date }
        effectiveDate: { type: string, format: date }
        partnerImpact: { type: string }
        calculationMethod: { type: string }
        amendmentLetterReference: { type: string }
        noticeChannels: { type: array, items: { type: string } }
        noticeMessage: { type: string }
        noticeMessageLanguage: { type: string }
    RevenueShare:
      type: object
      properties:
        id: { type: integer }
        contract_id: { type: integer }
        contract_title: { type: string }
        partner_name: { type: string }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        total_revenue: { type: number }
        share_percentage: { type: number }
        share_amount: { type: number }
        status: { type: string }
        paid_at: { type: string, format: date-time }
        notes: { type: string }
    RevenueShareInput:
      type: object
      required: [contractId, periodStart, periodEnd, totalRevenue, sharePercentage]
      properties:
        contractId: { type: integer }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        totalRevenue: { type: number }
        sharePercentage: { type: number }
        status: { type: string }
        notes: { type: string }
      description: >
        shareAmount is derived server-side from totalRevenue and
        sharePercentage; any client-supplied value is ignored.
    KPI:
      type: object
      properties:
        id: { type: integer }
        contract_id: { type: integer }
        contract_title: { type: string }
        name: { type: string }
        description: { type: string }
        target_value: { type: number }
        actual_value: { type: number }
        unit: { type: string }
        period_type: { type: string }
        status: { type: string }
    KPIInput:
      type: object
      required: [contractId, name, targetValue]
      properties:
        contractId: { type: integer }
        name: { type: string }
        description: { type: string }
        targetValue: { type: number }
        actualValue: { type: number }
        unit: { type: string }
        periodType: { type: string }
        status: { type: string }
    LegalDocument:
      type: object
      properties:
        id: { type: integer }
        contract_id: { type: integer }
        contract_title: { type: string }
        document_type: { type: string }
        document_name: { type: string }
        file_path: { type: string }
        file_url: { type: string }
        version: { type: string }
        status: { type: string }
    LegalDocumentInput:
      type: object
      required: [contractId, documentType, documentName]
      properties:
        contractId: { type: integer }
        documentType: { type: string }
        documentName: { type: string }
        filePath: { type: string }
        fileUrl: { type: string }
        version: { type: string }
        status: { type: string }
    Claim:
      type: object
      properties:
        id: { type: integer }
        tenant_id: { type: integer }
        partner_id: { type: integer }
        partner_name: { type: string }
        contract_id: { type: integer }
        contract_title: { type: string }
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        basis: { type: string }
        claimed_amount: { type: number }
        approved_amount: { type: number }
        currency: { type: string }
        status: { type: string, enum: [submitted, under_review, needs_clarification, approved, rejected] }
        decision_note: { type: string }
        reviewed_by: { type: integer }
        reviewed_at: { type: string, format: date-time }
    ClaimInput:
      type: object
      required: [partnerId]
      properties:
        partnerId: { type: integer }
        contractId: { type: integer }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        basis: { type: string }
        claimedAmount: { type: number }
        currency: { type: string }
    AuditEvent:
      type: object
      properties:
        id: { type: integer }
        tenant_id: { type: integer }
        actor_user_id: { type: integer }
        actor_name: { type: string }
        entity_type: { type: string }
        entity_id: { type: integer }
        action: { type: string }
        metadata: { type: object }
        created_at: { type: string, format: date-time }